Add password entry timestamps.
To facilitate timely password rotations, a date stamp will be
appended to any new or updated entry in the password database.
Anytime the password is retrieved, its updated time will be reported
to stderr. If the password has no timestamp (ie., it predates this
change), no updated time is shown.