Fix linger crash on OpenBSD.
Mishandling of the linger_flag caused the pledge() restrictions to
be wrong when linger was requested by environment variable. Later
when unlink() is called, pledge() will crash the process because
the cpath permission is missing.

Now linger_flag is set via environment variable by default, and can
be overridden by command line flag.